test: add a generic function for CLI kata-check command
Add a generic function to run CLI kata-check tests, shared by all the
args.
