+++ cli/kata-check_amd64_test.go
@@ -10,15 +10,12 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"path"
 	"path/filepath"
 	"regexp"
-	"strings"
 	"testing"
 
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
-	"github.com/urfave/cli"
 )
 
 func setupCheckHostIsVMContainerCapable(assert *assert.Assertions, cpuInfoFile string, cpuData []testCPUData, moduleData []testModuleData) {
@@ -47,36 +44,6 @@ func setupCheckHostIsVMContainerCapable(assert *assert.Assertions, cpuInfoFile s
 }
 
 func TestCCCheckCLIFunction(t *testing.T) {
-	assert := assert.New(t)
-
-	dir, err := ioutil.TempDir("", "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(dir)
-
-	_, config, err := makeRuntimeConfig(dir)
-	assert.NoError(err)
-
-	savedSysModuleDir := sysModuleDir
-	savedProcCPUInfo := procCPUInfo
-
-	cpuInfoFile := filepath.Join(dir, "cpuinfo")
-
-	// XXX: override
-	sysModuleDir = filepath.Join(dir, "sys/module")
-	procCPUInfo = cpuInfoFile
-
-	defer func() {
-		sysModuleDir = savedSysModuleDir
-		procCPUInfo = savedProcCPUInfo
-	}()
-
-	err = os.MkdirAll(sysModuleDir, testDirMode)
-	if err != nil {
-		t.Fatal(err)
-	}
-
 	var cpuData []testCPUData
 	var moduleData []testModuleData
 
@@ -94,50 +61,7 @@ func TestCCCheckCLIFunction(t *testing.T) {
 		moduleData = []testModuleData{}
 	}
 
-	devNull, err := os.OpenFile(os.DevNull, os.O_WRONLY, 0666)
-	assert.NoError(err)
-	defer devNull.Close()
-
-	savedLogOutput := kataLog.Logger.Out
-
-	// discard normal output
-	kataLog.Logger.Out = devNull
-
-	defer func() {
-		kataLog.Logger.Out = savedLogOutput
-	}()
-
-	setupCheckHostIsVMContainerCapable(assert, cpuInfoFile, cpuData, moduleData)
-
-	ctx := createCLIContext(nil)
-	ctx.App.Name = "foo"
-	ctx.App.Metadata["runtimeConfig"] = config
-
-	// create buffer to save logger output
-	buf := &bytes.Buffer{}
-
-	// capture output this time
-	kataLog.Logger.Out = buf
-
-	fn, ok := kataCheckCLICommand.Action.(func(context *cli.Context) error)
-	assert.True(ok)
-
-	err = fn(ctx)
-	assert.NoError(err)
-
-	output := buf.String()
-
-	for _, c := range cpuData {
-		assert.True(findAnchoredString(output, c.vendorID))
-		for _, flag := range strings.Fields(c.flags) {
-			assert.True(findAnchoredString(output, flag))
-		}
-	}
-
-	for _, m := range moduleData {
-		name := path.Base(m.path)
-		assert.True(findAnchoredString(output, name))
-	}
+	genericCheckCLIFunction(t, cpuData, moduleData)
 }
 
 func TestCheckCheckKernelModulesNoNesting(t *testing.T) {
