+++ cli/kata-check_arm64_test.go
@@ -6,128 +6,34 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
 	"os"
-	"path"
 	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/urfave/cli"
 )
 
-func setupCheckHostIsVMContainerCapable(assert *assert.Assertions, cpuInfoFile string, moduleData []testModuleData) {
+func setupCheckHostIsVMContainerCapable(assert *assert.Assertions, cpuInfoFile string, cpuData []testCPUData, moduleData []testModuleData) {
 	//For now, Arm64 only deal with module check
+	_ = cpuData
+
 	createModules(assert, cpuInfoFile, moduleData)
 
 	err := makeCPUInfoFile(cpuInfoFile, "", "")
 	assert.NoError(err)
-
 }
 
 func TestCCCheckCLIFunction(t *testing.T) {
-	assert := assert.New(t)
-
-	dir, err := ioutil.TempDir("", "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(dir)
-
-	_, config, err := makeRuntimeConfig(dir)
-	assert.NoError(err)
-
-	savedSysModuleDir := sysModuleDir
-	savedProcCPUInfo := procCPUInfo
-
-	cpuInfoFile := filepath.Join(dir, "cpuinfo")
-
-	// XXX: override
-	sysModuleDir = filepath.Join(dir, "sys/module")
-	procCPUInfo = cpuInfoFile
-
-	defer func() {
-		sysModuleDir = savedSysModuleDir
-		procCPUInfo = savedProcCPUInfo
-	}()
-
-	err = os.MkdirAll(sysModuleDir, testDirMode)
-	if err != nil {
-		t.Fatal(err)
-	}
-
+	var cpuData []testCPUData
 	moduleData := []testModuleData{
 		{filepath.Join(sysModuleDir, "kvm"), true, ""},
 		{filepath.Join(sysModuleDir, "vhost"), true, ""},
 		{filepath.Join(sysModuleDir, "vhost_net"), true, ""},
 	}
 
-	devNull, err := os.OpenFile(os.DevNull, os.O_WRONLY, 0666)
-	assert.NoError(err)
-	defer devNull.Close()
-
-	savedLogOutput := kataLog.Logger.Out
-
-	// discard normal output
-	kataLog.Logger.Out = devNull
-
-	defer func() {
-		kataLog.Logger.Out = savedLogOutput
-	}()
-
-	setupCheckHostIsVMContainerCapable(assert, cpuInfoFile, moduleData)
-
-	ctx := createCLIContext(nil)
-	ctx.App.Name = "foo"
-	ctx.App.Metadata["runtimeConfig"] = config
-
-	// create buffer to save logger output
-	buf := &bytes.Buffer{}
-
-	// capture output this time
-	kataLog.Logger.Out = buf
-
-	fn, ok := kataCheckCLICommand.Action.(func(context *cli.Context) error)
-	assert.True(ok)
-
-	err = fn(ctx)
-	assert.NoError(err)
-
-	output := buf.String()
-
-	for _, m := range moduleData {
-		name := path.Base(m.path)
-		assert.True(findAnchoredString(output, name))
-	}
-}
-
-func TestKvmIsUsable(t *testing.T) {
-	assert := assert.New(t)
-
-	dir, err := ioutil.TempDir("", "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(dir)
-
-	savedKvmDevice := kvmDevice
-	fakeKVMDevice := filepath.Join(dir, "kvm")
-	kvmDevice = fakeKVMDevice
-
-	defer func() {
-		kvmDevice = savedKvmDevice
-	}()
-
-	err = kvmIsUsable()
-	assert.Error(err)
-
-	err = createEmptyFile(fakeKVMDevice)
-	assert.NoError(err)
-
-	err = kvmIsUsable()
-	assert.Error(err)
+	genericCheckCLIFunction(t, cpuData, moduleData)
 }
 
 func TestGetCPUDetails(t *testing.T) {
