+++ cli/kata-check_ppc64le_test.go
@@ -6,17 +6,14 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
 	"os"
-	"path"
 	"path/filepath"
 	"testing"
 
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
-	"github.com/urfave/cli"
 )
 
 func setupCheckHostIsVMContainerCapable(assert *assert.Assertions, cpuInfoFile string, cpuData []testCPUData, moduleData []testModuleData) {
@@ -45,38 +42,8 @@ func setupCheckHostIsVMContainerCapable(assert *assert.Assertions, cpuInfoFile s
 }
 
 func TestCCCheckCLIFunction(t *testing.T) {
-	assert := assert.New(t)
-
-	dir, err := ioutil.TempDir("", "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(dir)
-
-	_, config, err := makeRuntimeConfig(dir)
-	assert.NoError(err)
-
-	savedSysModuleDir := sysModuleDir
-	savedProcCPUInfo := procCPUInfo
-
-	cpuInfoFile := filepath.Join(dir, "cpuinfo")
-
-	// XXX: override
-	sysModuleDir = filepath.Join(dir, "sys/module")
-	procCPUInfo = cpuInfoFile
-
-	defer func() {
-		sysModuleDir = savedSysModuleDir
-		procCPUInfo = savedProcCPUInfo
-	}()
-
-	err = os.MkdirAll(sysModuleDir, testDirMode)
-	if err != nil {
-		t.Fatal(err)
-	}
-
 	cpuData := []testCPUData{
-		{"", "", false},
+		fakeCPUData,
 	}
 
 	moduleData := []testModuleData{
@@ -84,43 +51,7 @@ func TestCCCheckCLIFunction(t *testing.T) {
 		{filepath.Join(sysModuleDir, "kvm_hv"), false, "Y"},
 	}
 
-	devNull, err := os.OpenFile(os.DevNull, os.O_WRONLY, 0666)
-	assert.NoError(err)
-	defer devNull.Close()
-
-	savedLogOutput := kataLog.Logger.Out
-
-	// discard normal output
-	kataLog.Logger.Out = devNull
-
-	defer func() {
-		kataLog.Logger.Out = savedLogOutput
-	}()
-
-	setupCheckHostIsVMContainerCapable(assert, cpuInfoFile, cpuData, moduleData)
-
-	ctx := createCLIContext(nil)
-	ctx.App.Name = "foo"
-	ctx.App.Metadata["runtimeConfig"] = config
-
-	// create buffer to save logger output
-	buf := &bytes.Buffer{}
-
-	// capture output this time
-	kataLog.Logger.Out = buf
-
-	fn, ok := kataCheckCLICommand.Action.(func(context *cli.Context) error)
-	assert.True(ok)
-
-	err = fn(ctx)
-	assert.NoError(err)
-
-	output := buf.String()
-
-	for _, m := range moduleData {
-		name := path.Base(m.path)
-		assert.True(findAnchoredString(output, name))
-	}
+	genericCheckCLIFunction(t, cpuData, moduleData)
 }
 
 func TestArchKernelParamHandler(t *testing.T) {
